test: test cases for new API { ok, nil, err }
