+++ index.js
@@ -1,1 +1,1 @@
-var C=(y)=>typeof y==="function";var g=(y)=>C(y)&&y.length<2,T=(y,x)=>Object.prototype.toString.call(y)===`[object ${x}]`,k=(y)=>(x)=>x instanceof y,I=k(Error),b=k(Promise),q=(y)=>I(y)?y:new Error(String(y)),p=(y,x)=>{if(!x)return!1;return y.name===x.name&&y.message===x.message};if(!("requestAnimationFrame"in globalThis))globalThis.requestAnimationFrame=(y)=>setTimeout(y,16);var Z,D=new Set,U=0,P=new Map,M,h=()=>{M=void 0;for(let y of P.values()){for(let x of y.values())x();y.clear()}},v=()=>{if(M)cancelAnimationFrame(M);M=requestAnimationFrame(h)};queueMicrotask(h);var F=(y)=>{if(Z&&!y.includes(Z))y.push(Z)},N=(y)=>{for(let x of y)U?D.add(x):x()},m=()=>{while(D.size){let y=Array.from(D);D.clear();for(let x of y)x()}},t=(y)=>{U++,y(),m(),U--},R=(y,x)=>{let $=Z;Z=x,y(),Z=$},s=(y,x)=>new Promise(($,L)=>{let B=()=>{try{$(y())}catch(z){L(z)}};if(x){let[z,G]=x;if(!P.has(z))P.set(z,new Map);P.get(z).set(G,B)}v()}),u=async()=>new Promise(requestAnimationFrame);function Y(y,...x){let $=C(y)?{ok:y}:y,L=()=>R(()=>{let B=E(x,$);if(I(B))throw new Error("Unhandled error in effect:",{cause:B})},L);L()}var d="Computed",j=(y,...x)=>{let $=C(y)?{ok:y}:y,L=[],B=K,z,G=!0,H=!1,J=!1,A=()=>{if(G=!0,!H)N(L)},Q=()=>R(()=>{if(J)throw new Error("Circular dependency detected");let V=(X)=>{if(!Object.is(X,B))B=X,G=!1,z=void 0,H=!1},o=(X)=>{let _=q(X);H=p(_,z),z=_};H=!0,J=!0;let W=E(x,$);if(b(W))W.then((X)=>{V(X),N(L)}).catch(o);else if(I(W))o(W);else if(W==null)H=K===B,B=K,z=void 0;else V(W);J=!1},A),S={[Symbol.toStringTag]:d,get:()=>{if(F(L),m(),G)Q();if(z)throw z;return B},map:(V)=>j(()=>V(S.get())),match:(V)=>Y(V,S)};return S},f=(y)=>T(y,d);var n="State",w=(y)=>{let x=[],$=y,L={[Symbol.toStringTag]:n,get:()=>{return F(x),$},set:(B)=>{if(Object.is($,B))return;if($=B,N(x),K===$)x.length=0},update:(B)=>{L.set(B($))},map:(B)=>{return j(()=>B(L.get()))},match:(B)=>Y(B,L)};return L},O=(y)=>T(y,n);var K=Symbol(),i=(y)=>O(y)||f(y),c=(y)=>i(y)?y:g(y)?j(y):w(y),E=(y,x)=>{let{ok:$,nil:L,err:B}=x,z=[],G=[],H=!1;for(let A of y)try{let Q=A.get();if(Q===K)H=!0;z.push(Q)}catch(Q){G.push(q(Q))}let J=void 0;try{if(!H&&!G.length)J=$(...z);else if(G.length&&B)J=B(...G);else if(H&&L)J=L()}catch(A){J=q(A)}finally{return J}};export{R as watch,c as toSignal,w as state,O as isState,i as isSignal,f as isComputed,s as enqueue,Y as effect,j as computed,t as batch,u as animationFrame,K as UNSET};
+var j=(y)=>typeof y==="function";var O=(y)=>j(y)&&y.length<2,N=(y,x)=>Object.prototype.toString.call(y)===`[object ${x}]`,g=(y)=>(x)=>x instanceof y,A=g(Error),o=g(Promise),C=(y)=>A(y)?y:new Error(String(y)),k=(y,x)=>{if(!x)return!1;return y.name===x.name&&y.message===x.message};var X,D=new Set,_=0,P=new Map,S,b=()=>{S=void 0;for(let y of P.values()){for(let x of y.values())x();y.clear()}},v=()=>{if(S)cancelAnimationFrame(S);S=requestAnimationFrame(b)};queueMicrotask(b);var M=(y)=>{if(X&&!y.includes(X))y.push(X)},I=(y)=>{for(let x of y)_?D.add(x):x()},U=()=>{while(D.size){let y=Array.from(D);D.clear();for(let x of y)x()}},t=(y)=>{_++,y(),U(),_--},q=(y,x)=>{let z=X;X=x,y(),X=z},u=(y,x)=>new Promise((z,L)=>{let B=()=>{try{z(y())}catch(G){L(G)}};if(x){let[G,J]=x;if(!P.has(G))P.set(G,new Map);P.get(G).set(J,B)}v()});function R(y,...x){let z=j(y)?{ok:y}:y,L=()=>q(()=>{let B=T(x,z);if(A(B))console.error("Unhandled error in effect:",B)},L);L()}var p="Computed",Z=(y,...x)=>{let z=j(y)?{ok:y}:y,L=[],B=Q,G,J=!0,K=!1,H=!1,W=($)=>{if(!Object.is($,B))B=$,J=!1,G=void 0,K=!1},V=()=>{K=Q===B,B=Q,G=void 0},m=($)=>{let Y=C($);K=k(Y,G),B=Q,G=Y},n=()=>{if(J=!0,!K)I(L)},i=()=>q(()=>{if(H)throw new Error("Circular dependency detected");K=!0,H=!0;let $=T(x,z);if(o($))V(),$.then((Y)=>{W(Y),I(L)}).catch(m);else if($==null||Q===$)V();else if(A($))m($);else W($);H=!1},n),F={[Symbol.toStringTag]:p,get:()=>{if(M(L),U(),J)i();if(G)throw G;return B},map:($)=>Z(()=>$(F.get())),match:($)=>R($,F)};return F},f=(y)=>N(y,p);var h="State",w=(y)=>{let x=[],z=y,L={[Symbol.toStringTag]:h,get:()=>{return M(x),z},set:(B)=>{if(Object.is(z,B))return;if(z=B,I(x),Q===z)x.length=0},update:(B)=>{L.set(B(z))},map:(B)=>{return Z(()=>B(L.get()))},match:(B)=>R(B,L)};return L},E=(y)=>N(y,h);var Q=Symbol(),d=(y)=>E(y)||f(y),c=(y)=>d(y)?y:O(y)?Z(y):w(y),T=(y,x)=>{let{ok:z,nil:L,err:B}=x,G=[],J=[],K=!1;for(let W of y)try{let V=W.get();if(V===Q)K=!0;G.push(V)}catch(V){J.push(C(V))}let H=void 0;try{if(K&&L)H=L();else if(J.length)H=B?B(...J):J[0];else if(!K)H=z(...G)}catch(W){if(H=C(W),B)H=B(H)}finally{return H}};export{q as watch,c as toSignal,w as state,E as isState,d as isSignal,f as isComputed,u as enqueue,R as effect,Z as computed,t as batch,Q as UNSET};
